+++ src/CarbonAware.CLI/test/integrationTests/Commands/Location/LocationCommandTests.cs
@@ -10,7 +10,6 @@ namespace CarbonAware.CLI.IntegrationTests.Commands.Location;
 [TestFixture(DataSourceType.JSON)]
 [TestFixture(DataSourceType.WattTime)]
 [TestFixture(DataSourceType.ElectricityMaps)]
-[TestFixture(DataSourceType.ElectricityMapsFree)]
 class LocationCommandTests : IntegrationTestingBase
 {
     public LocationCommandTests(DataSourceType dataSource) : base(dataSource) { }