+++ src/CarbonAware.DataSources/CarbonAware.DataSources.ElectricityMapsFree/src/Constants/Paths.cs
@@ -1,6 +1,0 @@
-﻿namespace CarbonAware.DataSources.ElectricityMapsFree.Constants;
-
-internal class Paths
-{
-    public const string Latest = "latest";
-}